+++ src/utils/commonUtils.ts
@@ -1,4 +1,4 @@
-import { Readable, Stream, Writable } from 'stream'
+import { Readable, Stream } from 'stream'
 
 /**
  * Converts the given number to a Buffer.
@@ -48,11 +48,6 @@ export const streamToData = (inputStream: Readable): Promise<Buffer> => new Prom
     .on('error', reject)
     .on('end', () => resolve(output))
 })
-
-export const writeInStream = async (stream: Writable, data: Buffer): Promise<void> => { // this should basically never be needed for crypto streams, but hey, better do things cleanly
-  const shouldWait = !stream.write(data)
-  if (shouldWait) await new Promise(resolve => stream.once('drain', resolve))
-}
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export const mergeInto = (x: Record<string, any>, y: Record<string, any>): void => {